main/src/requirements.txt
exif==1.3.5
fitz==0.0.1.dev2
haishoku==1.1.8
layoutparser==0.3.4
matlab==0.1
numpy==1.23.2
opencv_python==4.6.0.66
Pillow==9.3.0
protobuf==4.21.12
pyenchant==3.2.2
PyMuPDF==1.20.2
pytesseract==0.3.10
reportlab==3.6.11
wcag_contrast_ratio==0.9
